finished pagination on the plants show page
I added some nice tabs, it looks a lot better now. It also disables tabs that contain no information (so the user doesn't open a tab with nothing in it. I wrote a helper method and some JQuery for that. Check it out in app/views/current_plants/show.html.erb. I also made another attempt to stop git syncing the log and tmp files, so I hope that works.