trying to get nested form for photos to work. Archived big dataset for later use
I put the Bento data into a zip file (and saved it on my computer just in case) for later use as it's easier to test with a smaller number of plants. I'm trying to get the nested form working for photos but it's not working. It doesn't save a photo (the description of it, anyway) when you save a plant.