many page changes
The navbar has had a rework. Was annoying to do but it finally resizes properly with the screen. I've started adding a couple new variables for each page, 'prev_page' and 'prev_link', which hold the name of the previous page, and previous link, respectively. It basically tells the navbar what to put on the left hand side, to allow the user to navigate easier. Going to work on the bottom navigation part now.