fixed checkboxes for me
i think the problem was the height for them was really big, cos it was
taking the same height as form-controls, which was all the other form
input areas (like text areas and drop down menus as well). so i made a
new css element just for the checkboxes and set it to 15px. I think
this is just right - not too big and not too small - so that it looks
good on my phone and my mac (too big looked better on my mac but really
bad on my phone) i think 15px is good.